Login and Verfication Starting in Making an image for each Student making admin

+++ Delete.php
@@ -1,5 +1,10 @@
 <?php
+session_start();
 include "Database.php";
+if(empty($_SESSION['name'])){
+    header("location: index.php");
+    exit;
+}
 $db = new Database();
 $check = $db->delete($_GET['id']);
 header("location: index.php");
